fix: DagResource API works reliably for simple requests
Redesign agents and especially RedshiftCsv to be able to handle
requests without a restart. Now files are read every time.

Add a cache for parsed queries to reduce I/O and CPU for parsing.

Redesign Etl class to create DAGs only.

Improve code in analyzer and visitors to be more DRY. This was also
required to correctly handle queries that are not DML.